new file:   frontend/assets/MREC_logo.png 	modified:   frontend/lib/Faculty_Module/home_screen.dart 	modified:   frontend/lib/Faculty_Module/main_screen.dart 	modified:   frontend/lib/main.dart 	modified:   frontend/lib/models/performance_model.dart 	modified:   frontend/lib/models/user_model.dart 	modified:   frontend/lib/providers.dart/user_provider.dart 	modified:   frontend/lib/screens/attendance_screen.dart 	modified:   frontend/lib/screens/books_screen.dart 	modified:   frontend/lib/screens/home_screen.dart 	modified:   frontend/lib/screens/library_books_taken_screen.dart 	modified:   frontend/lib/screens/library_screen.dart 	modified:   frontend/lib/screens/login_screen.dart 	modified:   frontend/lib/screens/performance_screen.dart 	modified:   frontend/lib/screens/profile_screen.dart 	modified:   frontend/lib/screens/settings_screen.dart 	modified:   frontend/lib/screens/transport_screen.dart 	modified:   frontend/lib/services/auth.dart 	new file:   frontend/lib/services/ip.dart 	modified:   frontend/lib/widgets/book_item.dart 	modified:   frontend/lib/widgets/linear_progress_bar.dart 	new file:   frontend/lib/widgets/one_more_performance_card.dart 	modified:   frontend/lib/widgets/performance_card.dart 	modified:   frontend/pubspec.yaml 	modified:   server/controllers/userDataController.js 	modified:   server/models/userDetailsModel.js

+++ frontend/lib/models/performance_model.dart
@@ -5,7 +5,7 @@ class Performance {
   final String rollno;
   final int mid;
   final int mid_scored;
-  final int cgpa;
+  final double cgpa;
   final List<dynamic> previous_cgpa;
   final int backlogs;
   Performance({
@@ -33,7 +33,7 @@ class Performance {
       rollno: map['rollno'] as String,
       mid: map['mid'] as int,
       mid_scored: map['mid_scored'] as int,
-      cgpa: map['cgpa'] as int,
+      cgpa: map['cgpa'] as double,
       previous_cgpa:
           List<dynamic>.from((map['previous_cgpa'] as List<dynamic>)),
       backlogs: map['backlogs'] as int,
